# openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "SIMS",
    "version": "v1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Development Server"
    }
  ],
  "paths": {
    "/inventory/restock": {
      "put": {
        "tags": [
          "Restock"
        ],
        "operationId": "restock",
        "parameters": [
          {
            "name": "skunumber",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "quantity",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/API_Response"
                }
              }
            }
          }
        }
      }
    },
    "/inventory/item": {
      "put": {
        "tags": [
          "Inventory"
        ],
        "operationId": "updateAnItem",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/itemRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/API_Response"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Inventory"
        ],
        "operationId": "addNewItem",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/itemRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/API_Response"
                }
              }
            }
          }
        }
      }
    },
    "/inventory/addStock": {
      "put": {
        "tags": [
          "Restock"
        ],
        "operationId": "addStockForItem",
        "parameters": [
          {
            "name": "skunumber",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "quantity",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/API_Response"
                }
              }
            }
          }
        }
      }
    },
    "/account/revokeroles/{username}": {
      "put": {
        "tags": [
          "Account"
        ],
        "operationId": "revokeRolesOfAnAccount",
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "roles",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/API_Response"
                }
              }
            }
          }
        }
      }
    },
    "/account/revoke/{username}": {
      "put": {
        "tags": [
          "Account"
        ],
        "operationId": "removeAnAccount",
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/API_Response"
                }
              }
            }
          }
        }
      }
    },
    "/account/approve/{username}": {
      "put": {
        "tags": [
          "Account"
        ],
        "operationId": "approveAnAccount",
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/API_Response"
                }
              }
            }
          }
        }
      }
    },
    "/updateLastActive": {
      "post": {
        "tags": [
          "Account"
        ],
        "operationId": "logout",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/authenticationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/API_Response"
                }
              }
            }
          }
        }
      }
    },
    "/register": {
      "post": {
        "tags": [
          "Account"
        ],
        "operationId": "registerAccount",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/authenticationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/API_Response"
                }
              }
            }
          }
        }
      }
    },
    "/inventory/itemWithImage": {
      "post": {
        "tags": [
          "Inventory"
        ],
        "operationId": "addNewItemWithImage",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/itemWithImageRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/API_Response"
                }
              }
            }
          }
        }
      }
    },
    "/authenticate": {
      "post": {
        "tags": [
          "Account"
        ],
        "operationId": "returnJWTAfterAuthentication",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/authenticationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/API_Response"
                }
              }
            }
          }
        }
      }
    },
    "/account/addroles/{username}": {
      "post": {
        "tags": [
          "Account"
        ],
        "operationId": "updateRolesOfAnAccount",
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "roles",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/API_Response"
                }
              }
            }
          }
        }
      }
    },
    "/inventory/stock/all": {
      "get": {
        "tags": [
          "Restock"
        ],
        "operationId": "allStockInfo",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/API_Response"
                }
              }
            }
          }
        }
      }
    },
    "/inventory/item/sku/{SKU}": {
      "get": {
        "tags": [
          "Inventory"
        ],
        "operationId": "getAnItemUsingSKUNumber",
        "parameters": [
          {
            "name": "SKU",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/API_Response"
                }
              }
            }
          }
        }
      }
    },
    "/inventory/item/id/{ID}": {
      "get": {
        "tags": [
          "Inventory"
        ],
        "operationId": "getAnItemUsingID",
        "parameters": [
          {
            "name": "ID",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/API_Response"
                }
              }
            }
          }
        }
      }
    },
    "/inventory/demandforecasting/{skunumber}": {
      "get": {
        "tags": [
          "Inventory"
        ],
        "operationId": "getDemandForAnItem",
        "parameters": [
          {
            "name": "skunumber",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/API_Response"
                }
              }
            }
          }
        }
      }
    },
    "/inventory/category/all": {
      "get": {
        "tags": [
          "Inventory"
        ],
        "operationId": "getAllCategory",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/API_Response"
                }
              }
            }
          }
        }
      }
    },
    "/inventory/allbycategory": {
      "get": {
        "tags": [
          "Inventory"
        ],
        "operationId": "getFromInventoryByCategory",
        "parameters": [
          {
            "name": "pagenumber",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "pagesize",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "category",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "categoryid",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/API_Response"
                }
              }
            }
          }
        }
      }
    },
    "/inventory/all": {
      "get": {
        "tags": [
          "Inventory"
        ],
        "operationId": "getAllFromInventory",
        "parameters": [
          {
            "name": "pagenumber",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "pagesize",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/API_Response"
                }
              }
            }
          }
        }
      }
    },
    "/historicalsales/{year}": {
      "get": {
        "tags": [
          "Sales"
        ],
        "operationId": "getSalesForAYear",
        "parameters": [
          {
            "name": "year",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/API_Response"
                }
              }
            }
          }
        }
      }
    },
    "/analytics/all": {
      "get": {
        "tags": [
          "Dashboard"
        ],
        "operationId": "getAllAnalytics",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/API_Response"
                }
              }
            }
          }
        }
      }
    },
    "/account/details/{username}": {
      "get": {
        "tags": [
          "Account"
        ],
        "operationId": "getAccountDetails",
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/API_Response"
                }
              }
            }
          }
        }
      }
    },
    "/account/details/allActive": {
      "get": {
        "tags": [
          "Account"
        ],
        "operationId": "getAllAccountDetails",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/API_Response"
                }
              }
            }
          }
        }
      }
    },
    "/inventory/item/{skunumber}": {
      "delete": {
        "tags": [
          "Inventory"
        ],
        "operationId": "deleteAnItem",
        "parameters": [
          {
            "name": "skunumber",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/API_Response"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "API_Response": {
        "type": "object",
        "properties": {
          "httpStatus": {
            "type": "string",
            "enum": [
              "100 CONTINUE",
              "101 SWITCHING_PROTOCOLS",
              "102 PROCESSING",
              "103 CHECKPOINT",
              "200 OK",
              "201 CREATED",
              "202 ACCEPTED",
              "203 NON_AUTHORITATIVE_INFORMATION",
              "204 NO_CONTENT",
              "205 RESET_CONTENT",
              "206 PARTIAL_CONTENT",
              "207 MULTI_STATUS",
              "208 ALREADY_REPORTED",
              "226 IM_USED",
              "300 MULTIPLE_CHOICES",
              "301 MOVED_PERMANENTLY",
              "302 FOUND",
              "302 MOVED_TEMPORARILY",
              "303 SEE_OTHER",
              "304 NOT_MODIFIED",
              "305 USE_PROXY",
              "307 TEMPORARY_REDIRECT",
              "308 PERMANENT_REDIRECT",
              "400 BAD_REQUEST",
              "401 UNAUTHORIZED",
              "402 PAYMENT_REQUIRED",
              "403 FORBIDDEN",
              "404 NOT_FOUND",
              "405 METHOD_NOT_ALLOWED",
              "406 NOT_ACCEPTABLE",
              "407 PROXY_AUTHENTICATION_REQUIRED",
              "408 REQUEST_TIMEOUT",
              "409 CONFLICT",
              "410 GONE",
              "411 LENGTH_REQUIRED",
              "412 PRECONDITION_FAILED",
              "413 PAYLOAD_TOO_LARGE",
              "413 REQUEST_ENTITY_TOO_LARGE",
              "414 URI_TOO_LONG",
              "414 REQUEST_URI_TOO_LONG",
              "415 UNSUPPORTED_MEDIA_TYPE",
              "416 REQUESTED_RANGE_NOT_SATISFIABLE",
              "417 EXPECTATION_FAILED",
              "418 I_AM_A_TEAPOT",
              "419 INSUFFICIENT_SPACE_ON_RESOURCE",
              "420 METHOD_FAILURE",
              "421 DESTINATION_LOCKED",
              "422 UNPROCESSABLE_ENTITY",
              "423 LOCKED",
              "424 FAILED_DEPENDENCY",
              "425 TOO_EARLY",
              "426 UPGRADE_REQUIRED",
              "428 PRECONDITION_REQUIRED",
              "429 TOO_MANY_REQUESTS",
              "431 REQUEST_HEADER_FIELDS_TOO_LARGE",
              "451 UNAVAILABLE_FOR_LEGAL_REASONS",
              "500 INTERNAL_SERVER_ERROR",
              "501 NOT_IMPLEMENTED",
              "502 BAD_GATEWAY",
              "503 SERVICE_UNAVAILABLE",
              "504 GATEWAY_TIMEOUT",
              "505 HTTP_VERSION_NOT_SUPPORTED",
              "506 VARIANT_ALSO_NEGOTIATES",
              "507 INSUFFICIENT_STORAGE",
              "508 LOOP_DETECTED",
              "509 BANDWIDTH_LIMIT_EXCEEDED",
              "510 NOT_EXTENDED",
              "511 NETWORK_AUTHENTICATION_REQUIRED"
            ]
          },
          "message": {
            "type": "string"
          },
          "data": {
            "type": "object"
          }
        }
      },
      "itemRequest": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "imagePath": {
            "type": "string"
          },
          "itemName": {
            "type": "string"
          },
          "dimensions": {
            "type": "string"
          },
          "volume": {
            "type": "string"
          },
          "itemDescription": {
            "type": "string"
          },
          "unitPrice": {
            "type": "string"
          },
          "sku": {
            "type": "string"
          }
        }
      },
      "authenticationRequest": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        }
      },
      "itemWithImageRequest": {
        "type": "object",
        "properties": {
          "image": {
            "type": "string",
            "format": "binary"
          },
          "imagePath": {
            "type": "string"
          },
          "itemName": {
            "type": "string"
          },
          "dimensions": {
            "type": "string"
          },
          "volume": {
            "type": "string"
          },
          "itemDescription": {
            "type": "string"
          },
          "unitPrice": {
            "type": "string"
          },
          "sku": {
            "type": "string"
          }
        }
      }
    }
  }
}