# src/main/java/com/ebiggerr/sims/domain/request/authenticationRequest.java
package com.ebiggerr.sims.domain.request;

public class authenticationRequest {

    private String username;
    private String password;

    public authenticationRequest(String username, String password) {
        this.username = username;
        this.password = password;
    }

    public String getUsername() {
        return username;
    }

    public void setUsername(String username) {
        this.username = username;
    }

    public String getPassword() {
        return password;
    }

    public void setPassword(String password) {
        this.password = password;
    }


}